add: [Rust] add a solution for problem 52

+++ rust/src/new.rs
@@ -27,7 +27,7 @@ pub fn solve() {{
   let org = 0;
   println!(\"(original: {{}})\", org);
   input! {{
-    n: usize,
+    n: u64,
   }}
 
   let start = Instant::now();
